add EmptyEntityMetaOptionsService interface
